Implementacion de UseCase's completa
