+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/domain/repository/MovieRepository.kt
@@ -1,8 +1,10 @@
 package edu.dyds.movies.domain.repository
 
+import edu.dyds.movies.domain.entity.Movie
 import edu.dyds.movies.domain.entity.QualifiedMovie
+import edu.dyds.movies.domain.entity.RemoteMovie
 
 interface MovieRepository {
-    suspend fun getAllMovies() : List<QualifiedMovie>
-    suspend fun getMovieDetail(id: Int)
+    suspend fun getPopularMovies() : List<RemoteMovie>
+    suspend fun getMovieDetail(id: Int): Movie?
 }
